Move SidePaneLabel to other side pane components
